third_party/blink/web_tests/fast/peerconnection/rtc-transport.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
'use strict';

promise_test(async t => {
  const transport = new RtcTransport({
    name: 'transport',
    iceServers: [],
    iceControlling: true,
  });

  assert_equals(transport.getReceivedPackets().length, 0);
}, 'RtcTransport can be created');
</script>
